Replace the use of CommentFlag with custom view
Because `CommentFlag` doesn't work with a custom comment model
I replaced the view that creates the flag with one that just sends me
an email when there's a new flag. Which might actually be more useful.
Also had to update URLs and the flag form, in order to use that view.

For #98