Switch from Wikimedia maps to Stadia
Was getting 429 status errors (or, reportedly, 403) from the
maps.wikimedia.org map tiles. Either we're using too many, and/or
we need to specify a UserAgent, as per their Terms of Use. I can't
see how to add a User Agent using Leaflet, so for now switching to
the Stadia map tiles, which don't require a User Agent or API key.