Get API's EntryListView working again
There was an error related to the django_filters filtering, something
about "name" being the wrong argument.

I just made a new `get_queryset()` method for the view to do the
filtering, which seems to work instead.

So removed the django_filters module, as this was all it was used for.