Add Topic.diary_references_brief property
When trying to fetch a Topic in the API that has a lot of diary entries
as references (like ID 180, with 945 or so Entries) then the db on
Heroku was using too much memory and we got a 500 error.

Added the `Topic.diary_references_brief` property which is an alternative
to `Topic.diary_references` except (a) it only returns the `id`,
`title` and `diary_date` fields and (b) doesn't set an order. This
uses a quicker query, so will hopefully work on Heroku.