Fix CSS appearing in Wikipedia articles
We were removing the `<style>` tags in the clean phase, but that
left the stuff inside the tags. So the strip phase couldn't strip
the `<style>` tags out, and so the stuff inside them was still there.

Now we leave the `<style>` tags in the clean phase, so the strip
phase can get rid of them.

Fixes #300