Remove footnote links from Diary entry RSS feeds
Leave the <sup>1</sup> markers, and the footnotes themselves, but
remove the links around the markers, and the return links at the
end of each footnote.

Because they're no use in RSS feeds (as far as I can tell), and certainly
no use in the daily emails that we generate from them.