Fix error when an entry/letter has invalid references
When submitting an Entry or Letter whose `text` contains a link to
an Encyclopedia Topic that doesn't exist (an invalid numeric ID)
the models' `make_references()` methods failed with an error.

Now we silently ignore invalid IDs in the texts.

We could validate the text but I fear that might be too strict in some
ways – don't want to validate the content of a text field like that.
I might be wrong!

Fixes #822